refactor: Move Logger to separate package
