+++ src/main/java/lsd/format/JsonConverter.kt
@@ -3,6 +3,7 @@ package lsd.format
 import com.fasterxml.jackson.core.JsonProcessingException
 import com.fasterxml.jackson.core.type.TypeReference
 import lsd.format.json.createObjectMapper
+import lsd.format.log.log
 
 private val objectMapper = createObjectMapper()
