--- src/main/java/lsd/format/Logger.kt
@@ -1,8 +1,0 @@
-package lsd.format
-
-import org.slf4j.Logger
-import org.slf4j.LoggerFactory
-
-inline fun <reified T> T.log(): Logger = LoggerFactory.getLogger(T::class.java)
-
-fun log(): Logger = LoggerFactory.getLogger(Thread.currentThread().stackTrace[2].methodName)
