+++ src/main/java/lsd/format/json/ByteArraySerializer.kt
@@ -3,7 +3,7 @@ package lsd.format.json
 import com.fasterxml.jackson.core.JsonGenerator
 import com.fasterxml.jackson.databind.SerializerProvider
 import com.fasterxml.jackson.databind.ser.std.StdSerializer
-import lsd.format.log
+import lsd.format.log.log
 import lsd.format.xml.indentXml
 import java.io.IOException
